requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
Django==3.2.3
idna==2.10
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.5
